requirements.txt
base58==2.1.1
ecdsa==0.19.1
pycryptodome==3.22.0
six==1.17.0
